backend/requirements.txt
# Python dependencies

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-classic>=0.1.0

# Vector Database
chromadb>=0.4.0
langchain-chroma>=0.1.0

# Data processing
pandas>=2.0.0
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0

# Google Generative AI (for PDF processing)
google-generativeai>=0.3.0

# Utilities
python-multipart>=0.0.6

# Observability
langfuse>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
